// web/src/model/team-settings.ts
export interface TeamSettings {
  alleMuessenSchwimmen: boolean;
  minSchwimmerProTeam: number;
  maxSchwimmerProTeam: number;
  minMaleProTeam: number;
  minFemaleProTeam: number;
  minStartsProSchwimmer: number;
  maxStartsProSchwimmer: number;
  anzahlTeams: number;
  maxZeitspanneProStaffelSeconds: string;
}